Add files via upload
Had to fix a small error in the fetch_workbook_contents method of file_reader, because it was not sending DatabaseExcel to itself properly. All tests run and pass.

Made a small change in the get_data_from_excel method, removing the TypeError exception which was originally only in there as a workaround for times when the class couldn't send itself. This workaround is no longer needed.

Next, a test with a duplicate key will be added for 100% coverage, as currently one line of the class is covered. Then with that, the refactoring of the old create_connection long method in DatabaseExcel will be complete and 100% working.