Add files via upload
I thought there wasn't test data to check for duplicate keys, as line 154 of database_excel.py wasn't being covered during testing. But it turned out the method it was depending on wasn't returning a boolean value. That's fixed now, and all tests run and pass with 100% coverage of the class.