Add files via upload
Created get and set methods for each attribute in this class only. Tested and works. Next I'll do the same for chart_pie.py and chart_bar.py, and then work on merging them together to solve this bad smell of duplicate code. It's not exactly duplicated, but the three classes are doing something very similar and could be merged into one (which also makes it more expendable in the future too)