Add files via upload
Created a parent class, parent_chart.py and so I was able to remove all 62 duplicate lines of code from each of the three chart types, and improve the extendability of the system. In total 186 lines of similar code were able to be combined into one parent class, reducing the code to a 1/3 of the original size, and making it considerably easier to add a new chart.