Add files via upload
Removed duplicate lines of code in pie_chart and bar_chart such as "ChartClass.set_title(ChartClass, title)" that were repeated in both choices of chart type (e.g. bmi or birthday, gender or sales). Instead they were moved down before "if title" so if there was a matching chart type (again e.g. bmi or birthday etc) it will call the set methods for those attributes like title, labels etc in the chart class.

With the previous update:
Created a parent class, parent_chart.py and so I was able to remove all 62 duplicate lines of code from each of the three chart types, and improve the extendability of the system. In total 186 lines of similar code were able to be combined into one parent class, reducing the code to a 1/3 of the original size, and making it considerably easier to add a new chart.